rename example to cmd, move to a single .go file
